third_party/blink/web_tests/wpt_internal/prerender/restriction-media-microphone.https.html

<!DOCTYPE html>
<title>Access to the Microphone of the user media device is deferred</title>
<meta name="timeout" content="long">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

promise_test(async t => {
  const bc = new BroadcastChannel('test-channel');
  t.add_cleanup(_ => bc.close());

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });
  const url = `resources/media-devices-access.https.html?&audio=true`;
  window.open(url, '_blank', 'noopener');
  // According to spec, gUM will resolve only if the window has focus.
  window.focus();

  const result = await gotMessage;
  const expected = [
    {
      event: 'started waiting navigator.mediaDevices.getUserMedia',
      prerendering: true
    },
    {
      event: 'prerendering change',
      prerendering: false
    },
    {
      event: 'finished waiting navigator.mediaDevices.getUserMedia',
      prerendering: false
    },
  ];
  assert_equals(result.length, expected.length);
  for (let i = 0; i < result.length; i++) {
    assert_equals(result[i].event, expected[i].event, `event${i}`);
    assert_equals(result[i].prerendering, expected[i].prerendering,
      `prerendering${i}`);
  }
}, `the access to the Microphone of the user media should be deferred until the
    prerendered page is activated`);

</script>
</body>
